// ui/lib/api/schemas/llmCatalog.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * OpenAPI spec version: 0.0.1
 */
import type { LlmCatalogModel } from './llmCatalogModel';
import type { LlmProviderCatalog } from './llmProviderCatalog';

export interface LlmCatalog {
  localModels: LlmCatalogModel[];
  providers: LlmProviderCatalog[];
}