// ui/lib/api/schemas/llmProviderCatalog.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * OpenAPI spec version: 0.0.1
 */
import type { LlmCatalogModel } from './llmCatalogModel';
import type { LlmProviderCatalogStatus } from './llmProviderCatalogStatus';

export interface LlmProviderCatalog {
  /** @nullable */
  baseUrl?: string | null;
  /** @nullable */
  error?: string | null;
  hasApiKey: boolean;
  id: string;
  models: LlmCatalogModel[];
  name: string;
  requiresApiKey: boolean;
  requiresBaseUrl: boolean;
  status: LlmProviderCatalogStatus;
}